move TableDataSource and TableStyle to Models
