+++ lib/Map/LegendHelper.js
@@ -10,7 +10,7 @@ var ColorMap = require('../Map/ColorMap');
 var formatNumberWithCommas = require('../Core/formatNumberWithCommas');
 var Legend = require('../Map/Legend');
 var LegendUrl = require('../Map/LegendUrl');
-var TableStyle = require('../Map/TableStyle');
+var TableStyle = require('../Models/TableStyle');
 var VarType = require('../Map/VarType');
 
 var simplestats = require('simple-statistics');
