+++ lib/Models/CsvCatalogItem.js
@@ -16,8 +16,8 @@ var ModelError = require('./ModelError');
 var overrideProperty = require('../Core/overrideProperty');
 var proxyCatalogItemUrl = require('./proxyCatalogItemUrl');
 var readText = require('../Core/readText');
-var TableDataSource = require('../Map/TableDataSource');
-var TableStyle = require('../Map/TableStyle');
+var TableDataSource = require('../Models/TableDataSource');
+var TableStyle = require('../Models/TableStyle');
 
 /**
  * A {@link CatalogItem} representing CSV data.
