+++ lib/Models/TableDataSource.js
@@ -36,7 +36,7 @@ var LegendHelper = require('../Map/LegendHelper');
 var ModelError = require('../Models/ModelError');
 var RegionProviderList = require('../Map/RegionProviderList');
 var TableStructure = require('../Core/TableStructure');
-var TableStyle = require('../Map/TableStyle');
+var TableStyle = require('../Models/TableStyle');
 var VarType = require('../Map/VarType');
 
 // Used for region mapping
