+++ lib/Models/TableStyle.js
